Leave config within the gem.
Before this PR, much of the config that GeoConcerns relies on was copied into
the host application via Rails generators. This PR moves those configs from
generator templates into the gem's code base, where the Rails host app will
still be able to find and load them, but they will be less likely to conflict
with other gems, or with future upstream changes to Hyrax.

This is in compliance with the possible Hydra Plugin Guideline currently
being discussed here:

  https://github.com/projecthydra-labs/hydra_plugins_wg/issues/29

Closes #297.